+++ components/src/core/InfoListItem/InfoListItem.styles.tsx
@@ -8,6 +8,7 @@ export type InfoListItemClasses = {
     avatar?: string;
     divider?: string;
     icon?: string;
+    info?: string;
     listItemText?: string;
     rightComponent?: string;
     separator?: string;
@@ -28,6 +29,18 @@ const getIconColor = (props: InfoListItemProps): string => {
 
 const isWrapEnabled = (props: InfoListItemProps): boolean => props.wrapSubtitle || props.wrapTitle;
 
+const getIconAlignment = (props: InfoListItemProps): 'flex-end' | 'flex-start' | 'center' => {
+    switch (props.iconAlign) {
+        case 'left':
+            return 'flex-start';
+        case 'right':
+            return 'flex-end';
+        case 'center':
+        default:
+            return 'center';
+    }
+};
+
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
     createStyles({
@@ -44,9 +57,13 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
                           props.backgroundColor !== 'transparent'
                             ? color(props.backgroundColor)
                                   .darken(0.08)
-                                  .hex()
+                                  .rgb()
+                                  .string()
                             : 'rgba(0,0,0,0.08)'
                         : undefined,
+            },
+            '&:focus': {
+                outline: 'none',
             },
         },
         avatar: {
@@ -56,8 +73,8 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
         divider: {
             position: 'absolute',
             bottom: 0,
-            right: 0,
-            left: (props) => (props.divider === 'full' ? 0 : 72),
+            right: (props) => (theme.direction === 'rtl' ? (props.divider === 'full' ? 0 : 72) : 0),
+            left: (props) => (theme.direction === 'ltr' ? (props.divider === 'full' ? 0 : 72) : 0),
             zIndex: 0,
         },
         listItemText: {
@@ -65,6 +82,13 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
         },
         icon: {
             color: (props) => getIconColor(props),
+            justifyContent: (props) => getIconAlignment(props),
+            backgroundColor: 'transparent',
+        },
+        info: {
+            fontWeight: 400,
+            lineHeight: 1.3,
+            color: (props) => props.fontColor || 'inherit',
         },
         rightComponent: {
             flex: '0 0 auto',
@@ -98,5 +122,11 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             display: 'block',
             color: (props) => props.fontColor || 'inherit',
         },
+        flipIcon:
+            theme.direction === 'rtl'
+                ? {
+                      transform: 'scaleX(-1)',
+                  }
+                : {},
     })
 );