+++ demos/storybook/.storybook/config.js
@@ -1,13 +1,18 @@
 import React from 'react';
 import { addDecorator, addParameters } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
-import { blue as ReactTheme } from '@pxblue/react-themes';
-import { blueDark as ReactThemeDark } from '@pxblue/react-themes';
+import { createMuiTheme, jssPreset, MuiThemeProvider, StylesProvider } from '@material-ui/core/styles';
+import { blue as ReactTheme, blueDark as ReactThemeDark } from '@pxblue/react-themes';
 import 'typeface-open-sans';
 import { pxblueTheme } from '@pxblue/storybook-themes';
-import { useDarkMode } from 'storybook-dark-mode';
 import { CssBaseline } from '@material-ui/core';
+import { create } from 'jss';
+import rtl from 'jss-rtl';
+import addons, { useEffect, useState } from '@storybook/addons';
+import { DIR_CHANGE_EVENT, getDirection } from '@pxblue/storybook-rtl-addon';
+import { useDarkMode } from 'storybook-dark-mode/dist';
+
+const channel = addons.getChannel();
 
 const newViewports = {
     iPhone5: {
@@ -44,6 +49,7 @@ if (window.top.location.hostname === 'localhost') {
     pxblueTheme.brandImage = require('../assets/pxblue-react.svg');
 }
 
+// Only set theme inside storybook canvas.
 const themeInit = { dark: pxblueTheme, light: pxblueTheme, current: 'light' };
 window.localStorage.setItem('sb-addon-themes-3', JSON.stringify(themeInit));
 
@@ -67,14 +73,33 @@ addParameters({
     },
 });
 
-export const appliedTheme = createMuiTheme(ReactTheme);
-export const appliedThemeDark = createMuiTheme(ReactThemeDark);
+// Configure JSS
+const jss = create({ plugins: [...jssPreset().plugins, rtl()] });
 
-addDecorator((storyFn) => (
-    <MuiThemeProvider theme={useDarkMode() ? appliedThemeDark : appliedTheme}>
-        <CssBaseline />
-        <div className={'wrapper'}>{storyFn()}</div>
-    </MuiThemeProvider>
-));
+const appendDirection = (theme) => Object.assign(theme, { direction: getDirection() });
+
+addDecorator((storyFn) => {
+    const [lightTheme, setLightTheme] = useState(createMuiTheme(appendDirection(ReactTheme)));
+    const [darkTheme, setDarkTheme] = useState(createMuiTheme(appendDirection(ReactThemeDark)));
+
+    useEffect(() => {
+        channel.on(DIR_CHANGE_EVENT, updateDirection);
+        return () => channel.off(DIR_CHANGE_EVENT, updateDirection);
+    }, [channel]);
+
+    const updateDirection = () => {
+        setLightTheme(createMuiTheme(appendDirection(ReactTheme)));
+        setDarkTheme(createMuiTheme(appendDirection(ReactThemeDark)));
+    };
+
+    return (
+        <StylesProvider jss={jss}>
+            <MuiThemeProvider theme={useDarkMode() ? darkTheme : lightTheme}>
+                <CssBaseline />
+                <div className={'wrapper'}>{storyFn()}</div>
+            </MuiThemeProvider>
+        </StylesProvider>
+    );
+});
 
 addDecorator(withKnobs({ escapeHTML: false }));