+++ demos/storybook/.storybook/main.js
@@ -20,6 +20,7 @@ module.exports = {
         // @accessibility remove the following line and all other lines following @accessibility in the app to hide the a11y addon
         '@storybook/addon-a11y/register',
         '@storybook/addon-google-analytics/register',
+        '@pxblue/storybook-rtl-addon/register',
     ],
     webpackFinal: async (config) => {
         config.module.rules.push({