+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 
-## V4.0.2
+## v4.1.0
 
-- Updates several prop types to `ReactNode` to support wider range of input values.
-- Adds `info` prop to `<InfoListItem>` to support a third line of text.
+-   Fixes RTL support issues in the `<DrawerLayout>`/`<Drawer>`
+-   Fixes an issue with text wrapping in the `<InfoListItem>` subtitle
+-   Fixes an issue with hover color alpha channel in `<InfoListItem>`
+-   Adds `iconAlign` prop to `<InfoListItem>` to align icon left (default), center or right
+-   Adds optional prop to the `<DrawerHeader>` to add a bottom divider
+-   Adds `<DropdownToolbar>` component
+
+## v4.0.2
+
+-   Updates several prop types to `ReactNode` to support wider range of input values.
+-   Adds `info` prop to `<InfoListItem>` to support a third line of text.
 
 ## v4.0.1
 
@@ -16,6 +25,7 @@
 -   Changed component font colors to address potential accessibility issues.
 
 **Breaking Changes:**
+
 -   Theme usage updated to Material UI v4+ format (requires v5+ of @pxblue/react-themes)
 
 ## v3.0.3
@@ -32,7 +42,7 @@
 
 -   A few props got renamed to avoid further ambiguities:
     -   `<DrawerNavGroup>` prop `content` has been renamed to `titleContent`.
-    -   Anything controling the look of a `NavItem` / `NestedNavItem`, has been renamed to include 'item' in them. These are inheritable properties that might get confusing in different drawer hierarchy levels. List of affected props:
+    -   Anything controlling the look of a `NavItem` / `NestedNavItem`, has been renamed to include 'item' in them. These are inheritable properties that might get confusing in different drawer hierarchy levels. List of affected props:
 
 | Previous              | Current                   |
 | --------------------- | ------------------------- |